requirements.txt
tabula-py>=1.3.1
